refactor(main): extract window state and context menu helpers
Move window-bounds cache (WindowBounds, flushWindowState, saveWindowBounds, getSavedBounds, getWindowOptions) into src/main/lib/windowState.ts and the context menu attach helper into src/main/lib/contextMenu.ts. Trim historical narration around the Tray docking import, the removed TRAY_ICON, and the chooser host pill text.

Amp-Thread-ID: https://ampcode.com/threads/T-019e1f94-0238-767d-8770-7db59757c516
