+++ src/main/lib/contextMenu.ts
@@ -1,0 +1,46 @@
+import { Menu, clipboard, shell } from 'electron'
+import type { BrowserWindow } from 'electron'
+import * as i18n from './i18n'
+
+export function attachContextMenu(
+  comfyWindow: BrowserWindow,
+  webContents?: Electron.WebContents,
+): void {
+  ;(webContents || comfyWindow.webContents).on('context-menu', (_event, params) => {
+    const { editFlags, isEditable, selectionText, linkURL } = params
+    const hasSelection = selectionText.trim().length > 0
+    const hasLink = linkURL.length > 0
+
+    if (!isEditable && !hasSelection && !hasLink) return
+
+    const menuItems: Electron.MenuItemConstructorOptions[] = []
+
+    if (hasLink) {
+      menuItems.push(
+        { label: i18n.t('contextMenu.openLinkInBrowser'), click: () => shell.openExternal(linkURL) },
+        { label: i18n.t('contextMenu.copyLinkAddress'), click: () => clipboard.writeText(linkURL) },
+      )
+    }
+
+    if (hasLink && (isEditable || hasSelection)) {
+      menuItems.push({ type: 'separator' })
+    }
+
+    if (isEditable) {
+      menuItems.push(
+        { label: i18n.t('contextMenu.cut'), role: 'cut', enabled: editFlags.canCut },
+        { label: i18n.t('contextMenu.copy'), role: 'copy', enabled: editFlags.canCopy },
+        { label: i18n.t('contextMenu.paste'), role: 'paste', enabled: editFlags.canPaste },
+        { type: 'separator' },
+        { label: i18n.t('contextMenu.selectAll'), role: 'selectAll', enabled: editFlags.canSelectAll },
+      )
+    } else if (hasSelection) {
+      menuItems.push(
+        { label: i18n.t('contextMenu.copy'), role: 'copy', enabled: editFlags.canCopy },
+        { label: i18n.t('contextMenu.selectAll'), role: 'selectAll', enabled: editFlags.canSelectAll },
+      )
+    }
+
+    Menu.buildFromTemplate(menuItems).popup({ window: comfyWindow })
+  })
+}
